pyproject.toml
[project]
name = "admin-cervice"
version = "0.1.0"
description = ""
authors = [
    {name = "colosss",email = "kocnevs.jk@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "grpcio (>=1.76.0,<2.0.0)",
    "grpcio-tools (>=1.76.0,<2.0.0)",
    "aiokafka (>=0.12.0,<0.13.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
